+++ runtime/calamari/src/lib.rs
@@ -907,8 +907,7 @@ impl_runtime_apis! {
 			list_benchmarks!(list, extra);
 
 			let storage_info = AllPalletsReversedWithSystemFirst::storage_info();
-
-			return (list, storage_info)
+			(list, storage_info)
 		}
 
 		fn dispatch_benchmark(